Returning a backdrop solution without using portals.
