+++ packages/client/src/components/NavMenu/sub-components/profile-menu.js
@@ -208,6 +208,7 @@ class ProfileMenu extends React.Component {
         open={open}
         clickOutsideAction={clickOutsideAction}
         forwardedRef={forwardedRef}
+        isDefaultMode={false}
         withBlur={true}
         isBannerVisible={isBannerVisible}
         withPortal={true}
@@ -236,7 +237,15 @@ class ProfileMenu extends React.Component {
   };
 
   render() {
-    return <StyledWrapper>{this.renderDropDown()}</StyledWrapper>;
+    const { open } = this.props;
+
+    const element = this.renderDropDown();
+
+    const root = document.getElementById("root");
+
+    const wrapper = <StyledWrapper>{element}</StyledWrapper>;
+
+    return <>{<Portal element={wrapper} appendTo={root} visible={open} />}</>;
   }
 }
