+++ packages/login/src/components/SimpleNav.tsx
@@ -34,6 +34,7 @@ import { mobile } from "@docspace/shared/utils/device";
 import { getLogoUrl } from "@docspace/shared/utils/common";
 import { Base, Dark } from "@docspace/shared/themes";
 import { ThemeKeys, WhiteLabelLogoType } from "@docspace/shared/enums";
+import LanguageComboboxWrapper from "./LanguageCombobox";
 
 const StyledSimpleNav = styled.div`
   display: none;
@@ -50,6 +51,12 @@ const StyledSimpleNav = styled.div`
 
   @media ${mobile} {
     display: flex;
+
+    .language-combo-box {
+      position: absolute;
+      top: 7px;
+      right: 8px;
+    }
   }
 `;
 
@@ -66,6 +73,7 @@ const SimpleNav = ({ systemTheme }: SimpleNavProps) => {
   return (
     <StyledSimpleNav id="login-header" theme={isDark ? Dark : Base}>
       <img src={logoUrl} alt="logo-url" />
+      <LanguageComboboxWrapper />
     </StyledSimpleNav>
   );
 };
